[00119] Add shared copyToClipboard helper with execCommand fallback and update all call sites
